relocate helper methods
