+++ packages/transaction-controller/src/gas-flows/OptimismLayer1GasFeeFlow.test.ts
@@ -1,12 +1,9 @@
-import type { TypedTransaction } from '@ethereumjs/tx';
-import { toBuffer } from '@ethereumjs/util';
 import { Contract } from '@ethersproject/contracts';
 import type { Provider } from '@metamask/network-controller';
 
 import { CHAIN_IDS } from '../constants';
 import type { Layer1GasFeeFlowRequest, TransactionMeta } from '../types';
 import { TransactionStatus } from '../types';
-import { buildUnserializedTransaction } from '../utils/layer1-gas-fee-flow';
 import { OptimismLayer1GasFeeFlow } from './OptimismLayer1GasFeeFlow';
 
 jest.mock('@ethersproject/contracts', () => ({
@@ -29,14 +26,9 @@ const TRANSACTION_META_MOCK: TransactionMeta = {
   },
 };
 const OPTIMISIM_LAYER_1_GAS_FEE_RESPONSE_MOCK = '0x123';
-const TRANSACTION_SERIALIZED_BUFFER_MOCK = toBuffer('0x987654321');
 
 describe('OptimismLayer1GasFeeFlow', () => {
   const contractMock = jest.mocked(Contract);
-  const buildUnserializedTransactionMock = jest.mocked(
-    buildUnserializedTransaction,
-  );
-  const serializeMock: jest.MockedFn<() => Buffer> = jest.fn();
   const contractGetL1FeeMock: jest.MockedFn<
     () => Promise<{ toHexString: () => string }>
   > = jest.fn();
@@ -54,11 +46,6 @@ describe('OptimismLayer1GasFeeFlow', () => {
     contractMock.mockReturnValue({
       getL1Fee: contractGetL1FeeMock,
     } as unknown as Contract);
-
-    serializeMock.mockReturnValue(TRANSACTION_SERIALIZED_BUFFER_MOCK);
-    buildUnserializedTransactionMock.mockReturnValue({
-      serialize: serializeMock,
-    } as unknown as TypedTransaction);
   });
 
   describe('matchesTransaction', () => {
@@ -82,13 +69,7 @@ describe('OptimismLayer1GasFeeFlow', () => {
       const flow = new OptimismLayer1GasFeeFlow();
       const response = await flow.getLayer1Fee(request);
 
-      expect(buildUnserializedTransactionMock).toHaveBeenCalledWith(
-        TRANSACTION_META_MOCK,
-      );
-
-      expect(contractGetL1FeeMock).toHaveBeenCalledWith(
-        TRANSACTION_SERIALIZED_BUFFER_MOCK,
-      );
+      expect(contractGetL1FeeMock).toHaveBeenCalledTimes(1);
 
       expect(response).toStrictEqual({
         layer1Fee: OPTIMISIM_LAYER_1_GAS_FEE_RESPONSE_MOCK,
