+++ packages/transaction-controller/src/utils/layer1-gas-fee-flow.test.ts
@@ -1,16 +1,12 @@
 import type { Provider } from '@metamask/network-controller';
 import type { Hex } from '@metamask/utils';
-import BN from 'bn.js';
 
 import {
   TransactionStatus,
   type Layer1GasFeeFlow,
   type TransactionMeta,
 } from '../types';
-import {
-  updateTransactionLayer1GasFee,
-  buildUnserializedTransaction,
-} from './layer1-gas-fee-flow';
+import { updateTransactionLayer1GasFee } from './layer1-gas-fee-flow';
 
 jest.mock('@metamask/controller-utils', () => ({
   ...jest.requireActual('@metamask/controller-utils'),
@@ -126,30 +122,6 @@ describe('updateTransactionLayer1GasFee', () => {
       });
 
       expect(unmatchingLayer1GasFeeFlow.getLayer1Fee).not.toHaveBeenCalled();
-    });
-  });
-});
-
-describe('buildUnserializedTransaction', () => {
-  it('returns a transaction that can be serialized and fed to an Optimism smart contract', () => {
-    const unserializedTransaction = buildUnserializedTransaction({
-      txParams: {
-        nonce: '0x0',
-        gasPrice: `0x${new BN('100').toString(16)}`,
-        gas: `0x${new BN('21000').toString(16)}`,
-        to: '0x0000000000000000000000000000000000000000',
-        value: `0x${new BN('10000000000000').toString(16)}`,
-        data: '0x0',
-      },
-      chainId: '0x1',
-    } as unknown as TransactionMeta);
-    expect(unserializedTransaction.toJSON()).toMatchObject({
-      nonce: '0x0',
-      gasPrice: '0x64',
-      gasLimit: '0x5208',
-      to: '0x0000000000000000000000000000000000000000',
-      value: '0x9184e72a000',
-      data: '0x00',
     });
   });
 });
